fix: adapt agent-args placeholder validation to current architecture, split tests
