+++ src/routines/command.rs
@@ -561,3 +561,7 @@ pub(crate) fn build_routine_command(routine: &Routine, agent: &AgentCommand) -> 
 #[cfg(test)]
 #[path = "command_tests.rs"]
 mod command_tests;
+
+#[cfg(test)]
+#[path = "command_placeholder_tests.rs"]
+mod command_placeholder_tests;
